TG-53 #in-progress Admin/user roles

+++ pik-shop-backend/src/main/java/pl/elka/pw/pik/shop/security/service/LogoutHandler.java
@@ -4,6 +4,7 @@ import org.springframework.security.core.Authentication;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 
 public class LogoutHandler implements org.springframework.security.web.authentication.logout.LogoutHandler {
 
@@ -15,6 +16,15 @@ public class LogoutHandler implements org.springframework.security.web.authentic
 
     @Override
     public void logout(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
+        if ("OPTIONS".equals(request.getMethod())) {
+            try {
+                response.getWriter().print("OK");
+                response.getWriter().flush();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            return;
+        }
         tokenAuthenticationService.removeAuthentication(request, response);
     }
 }
